Moved chat event handler to own class
